# All around optimizations.