--- a/include/ada/helpers.h
+++ b/include/ada/helpers.h
@@ -5,13 +5,12 @@
 #include "common_defs.h"
 
 #include <string_view>
-#include <vector>
 #include <optional>
 
 namespace ada::helpers {
 
-  std::vector<std::string> split_string_view(std::string_view input, char delimiter, bool skip_empty = true);
-  ada_really_inline uint64_t string_to_uint64(std::string_view view);
+  // return 256^x, might overflow. Valid inputs are 0,1,...,7.
+  constexpr uint64_t pow256(uint64_t x) { return uint64_t(1) << (8*x); }
   ada_really_inline std::optional<std::string_view> prune_fragment(std::string_view& input) noexcept;
 } // namespace ada::helpers