--- a/include/ada/parser.h
+++ b/include/ada/parser.h
@@ -8,8 +8,8 @@
 #include <string_view>
 
 namespace ada::parser {
-
-  std::optional<std::string> to_ascii(std::string_view plain, bool be_strict) noexcept;
+  // first_percent should be  = plain.find('%')
+  std::optional<std::string> to_ascii(std::string_view plain, bool be_strict, size_t first_percent);
 
   std::optional<uint64_t> parse_ipv4_number(std::string_view input);
   std::optional<ada::url_host> parse_opaque_host(std::string_view input);