--- a/include/ada/unicode.h
+++ b/include/ada/unicode.h
@@ -8,6 +8,7 @@ namespace ada::unicode {
 
   ada_really_inline constexpr bool is_forbidden_host_code_point(const char c) noexcept;
   ada_really_inline constexpr bool is_forbidden_domain_code_point(const char c) noexcept;
+  ada_really_inline constexpr bool is_alnum_plus(const char c) noexcept;
   ada_really_inline constexpr bool is_ascii_hex_digit(const char c) noexcept;
   ada_really_inline constexpr bool is_c0_control_or_space(const char c) noexcept;
   ada_really_inline constexpr bool is_ascii_tab_or_newline(const char c) noexcept;
@@ -18,7 +19,8 @@ namespace ada::unicode {
 
   // todo: these functions would be faster as noexcept maybe, but it could be unsafe since
   // they are allocating.
-  std::string percent_decode(const std::string_view input);
+  // first_percent should be  = plain.find('%')
+  std::string percent_decode(const std::string_view input, size_t first_percent);
   std::string percent_encode(const std::string_view input, const uint8_t character_set[]);
   void percent_encode_character(const char input, const uint8_t character_set[], std::string& out);