This is part of a larger strategy which consists in pulling up some work so that later steps are cheap. Ideally, you'd have an integrated first step which does a bunch of work quickly... identify the location of the first #, check whether it is all ASCII and so forth.

```
GCC12
Linux ip-172-31-33-234 5.15.0-1022-aws #26-Ubuntu SMP Thu Oct 13 12:59:25 UTC 2022 x86_64 x86_64 x86_64 GNU/Linux
Main branch:
BasicBench_AdaURL      13879 ns        13800 ns        59962 time/byte=17.0797ns
This PR:
BasicBench_AdaURL       8330 ns         8330 ns        85169 time/byte=10.3091ns
```

Your numbers will vary, especially as the benchmark may vary from one commit to the other, and the 'main branch' is in flux. However, this code is simpler and that alone is a significant gain. We got rid entirely of `std::vector` !!!